Allow python client to read the ip address
The IP Address is part of the system fields and the python client will not read it without adding the system fields to LinkFields in form.json. The entry objects are set to use the system fields, but never find the IP Address because the field is not included in the form's field definition.